Changed files packetSSL, receiver.cpp. Add file commandData.h
