+++ LARCmaCS/commandData.h
@@ -1,0 +1,13 @@
+#include "packetSSL.h"
+
+struct CommandData {
+    quint32 speed_x;
+    quint32 speed_y;
+    quint32 speed_r;
+    quint32 speed_dribbler;
+    quint8 dribbler_enable;
+    quint32 kicker_volatage_level;
+    quint8 kicker_charge_enable;
+    quint8 kick_up;
+    quint8 kick_forward;
+};
