+++ LARCmaCS/packetSSL.h
@@ -10,19 +10,9 @@
 
 struct PacketSSL
 {
-    PacketSSL()
-    {
-        memset(robots_blue, 0, TEAM_COUNT_d);
-        memset(robots_yellow, 0, TEAM_COUNT_d);
-        memset(balls, 0, BALL_COUNT_d);
-    }
+    PacketSSL();
 
-    PacketSSL(const PacketSSL &ps)
-    {
-        memcpy(robots_blue, ps.robots_blue, TEAM_COUNT_d);
-        memcpy(robots_yellow, ps.robots_yellow, TEAM_COUNT_d);
-        memcpy(balls, ps.balls, BALL_COUNT_d);
-    }
+    PacketSSL(const PacketSSL &ps);
 
     double robots_blue[48];
     double robots_yellow[48];
